requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.0
pymongo==4.6.1
dnspython==2.4.2
twilio==9.0.4
gunicorn==21.2.0
scikit-learn>=1.0.0
joblib>=1.0.0
numpy>=1.20.0
